+++ packages/admin-e2e-tests/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 -   Add test cases for the home screen tasklist and activity panels. #7509
-
+-   Update analytics overview tests to allow re-running the tests.
 # 0.1.0
 
 -   Released initial package